Migrate type and site handling
