+++ specimen/src/org/labkey/specimen/actions/SpecimenApiController.java
@@ -30,14 +30,12 @@ import org.labkey.api.security.RequiresPermission;
 import org.labkey.api.security.User;
 import org.labkey.api.security.UserManager;
 import org.labkey.api.security.permissions.ReadPermission;
-import org.labkey.api.specimen.SpecimenManagerNew;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.location.LocationImpl;
 import org.labkey.api.specimen.location.LocationManager;
-import org.labkey.api.specimen.model.PrimaryType;
+import org.labkey.specimen.model.PrimaryType;
 import org.labkey.api.specimen.model.SpecimenTypeSummary;
 import org.labkey.api.specimen.security.permissions.RequestSpecimensPermission;
-import org.labkey.specimen.settings.RepositorySettings;
 import org.labkey.api.study.StudyUtils;
 import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.NotFoundException;
@@ -52,6 +50,7 @@ import org.labkey.specimen.model.AdditiveType;
 import org.labkey.specimen.model.DerivativeType;
 import org.labkey.specimen.requirements.SpecimenRequest;
 import org.labkey.specimen.security.permissions.ManageRequestsPermission;
+import org.labkey.specimen.settings.RepositorySettings;
 import org.labkey.specimen.settings.SettingsManager;
 import org.springframework.validation.BindException;
 
@@ -643,7 +642,7 @@ public class SpecimenApiController extends SpringActionController
         public ApiResponse execute(SpecimenApiForm form, BindException errors)
         {
             Container container = form.getViewContext().getContainer();
-            SpecimenTypeSummary summary = SpecimenManagerNew.get().getSpecimenTypeSummary(container, getUser());
+            SpecimenTypeSummary summary = SpecimenManager.get().getSpecimenTypeSummary(container, getUser());
             final Map<String, Object> response = new HashMap<>();
 
             List<Map<String, Object>> primaryTypes = new ArrayList<>();
